Move backend routes to admin namespace
