+++ app/views/admin/calls/new.html.erb
@@ -1,0 +1,3 @@
+<h1 class="heading">New Call</h1>
+
+<%= render 'form', call: @call, url: admin_round_calls_path(@call.round) %>
