+++ app/views/admin/rounds/index.html.erb
@@ -1,0 +1,16 @@
+<h1 class="heading">Rounds</h1>
+
+<% @rounds.each do |round| %>
+  <div class="row">
+    <%= link_to round.name, [:admin, round], class: "row-title row-title-link p-3" %>
+
+    <div>
+      <%= link_to 'Edit', edit_admin_round_path(round), class: "row-button row-button-edit" %>
+      <%= link_to 'Delete', [:admin, round], method: :delete, data: { confirm: 'Are you sure?' }, class: "row-button row-button-delete" %>
+    </div>
+  </div>
+<% end %>
+
+<div class="row row-muted">
+  <%= link_to "+ New Round", new_admin_round_path, class: "row-title" %>
+</div>
