--- app/views/calls/new.html.erb
@@ -1,3 +1,0 @@
-<h1 class="heading">New Call</h1>
-
-<%= render 'form', call: @call, url: round_calls_path(@call.round) %>
