+++ app/views/layouts/application.html.erb
@@ -13,7 +13,7 @@
   <body class="bg-gray-200">
     <div class="bg-white mb-4 shadow-lg">
       <div class="max-w-screen-lg mx-auto p-3">
-        <%= link_to "Bingo", rounds_path %>
+        <%= link_to "Bingo", admin_rounds_path %>
       </div>
     </div>
