+++ spec/system/manage_calls_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.describe "ManageCalls", type: :system do
   it "allows you to create calls" do
     round = Round.create(name: "Test Round")
-    visit round_path(round)
+    visit admin_round_path(round)
     click_on "New Call"
     fill_in "Name", with: "Red Ute"
     click_on "Create Call"
@@ -13,17 +13,17 @@ RSpec.describe "ManageCalls", type: :system do
   it "shows calls on round source screen" do
     round = Round.create(name: "Test Round")
     Call.create(name: "Red Ute", round: round, hidden: false)
-    visit screen_round_path(round)
+    visit screen_admin_round_path(round)
     wait_for_connection
     expect(page).to have_content("Red Ute")
   end
 
   it "doesn't add new calls to source screen automatically" do
     round = Round.create(name: "Test Round")
-    visit round_path(round)
+    visit admin_round_path(round)
 
     using_session "screen" do
-      visit screen_round_path(round)
+      visit screen_admin_round_path(round)
       wait_for_connection
       expect(page).not_to have_content "Red Ute"
     end
@@ -40,10 +40,10 @@ RSpec.describe "ManageCalls", type: :system do
   it "can show and hide calls" do
     round = Round.create(name: "Test Round")
     call = Call.create(name: "Red Ute", round: round)
-    visit round_path(round)
+    visit admin_round_path(round)
 
     using_session "screen" do
-      visit screen_round_path(round)
+      visit screen_admin_round_path(round)
       wait_for_connection
       expect(page).not_to have_content "Red Ute"
     end
